Fullstack Open | part4 | 4.15 - 4.23 add tests to blog, add token auth, fix test routes

+++ part4/blog-app/controllers/blogRouter.js
@@ -1,8 +1,13 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blogs')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
 blogRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({})
+        .populate('user', { username: 1 })
     
     response.status(200).json(blogs)
 })
@@ -14,30 +19,55 @@ blogRouter.get('/:id', async (request, response) => {
     response.status(200).json(blog)
 })
 
-blogRouter.post('/', async (request, response) => {
+blogRouter.post('/', userExtractor, async (request, response) => {
     const blog = new Blog(request.body)
+    const token = request.token
+    const signedInUser = await User.findById(request.user)
+
+    if (!token) {
+        return response.status(401).json(
+            { error: "Missing or invalid token"}
+        )
+    }
 
     if (!blog.likes) {
         blog.likes = 0
     }
     
-    const result = await blog.save()
-    response.status(201).json(result.toJSON())
+    const user = await User.findById(signedInUser._id)
+
+    blog.user = user._id
+
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+
+    await User.findByIdAndUpdate(user._id, user, { new: true })
+
+    response.status(201).json(savedBlog.toJSON())
 })
 
-blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+blogRouter.delete('/:id', userExtractor, async (request, response) => {
+   const signedInUser = await User.findById(request.user)
+   const blog = await Blog.findById(request.params.id)
+
+   if (!blog.user || !(blog.user.toString() === signedInUser._id.toString())) {
+        return response.status(401).json({
+            error: "Only the user who created the blog can delete it"
+        })
+   }
+   
+   await Blog.findByIdAndRemove(request.params.id)
+   response.status(204).end()
 })
 
 blogRouter.put('/:id', async (request, response) => {
-    const body = request.body
+    const { title, author, url, likes } = request.body
 
     const updatedBlog = {
-        title: body.title,
-        author: body.author,
-        url: body.url,
-        likes: body.likes
+        title,
+        author,
+        url,
+        likes,
     }
 
     await Blog.findByIdAndUpdate(request.params.id, updatedBlog)
